pyproject.toml
[project]
name = "lab-wizard-repo"
version = "0.1.0"
description = "Main SNSPD wizard app"
requires-python = ">=3.14"
dependencies = ["lab-wizard", "ruamel-yaml>=0.18.16"]


[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["lab_wizard"]

[tool.uv]
package = true

[project.scripts]
wizard = "lab_wizard.wizard.cli:main"

[tool.uv.workspace]
members = ["lab_wizard"]

[tool.uv.sources]
lab-wizard = { workspace = true, editable = true }
